backend/requirements.txt
# AI Healthcare Chatbot Backend Requirements

# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database
pymongo==4.5.0
mongoengine==0.27.0

# AI/ML Models
langchain==0.1.0
langchain-community==0.0.10
langchain-groq==0.1.0
transformers==4.34.0
torch==2.11.0
numpy==1.24.3
scikit-learn==1.3.2

# Encryption
cryptography==41.0.5

# Environment
python-dotenv==1.0.0

# HTTP Client
httpx==0.25.0
requests==2.31.0

# Data Processing
pandas==2.1.3
pydantic==2.5.0

# Utilities
python-jose==3.3.0
pydantic-settings==2.1.0
